Add support for collating all
Relates to issue #4

Added a function called get_files_to_collate that takes a filename and
whether you want to collate all or not. This returns which filenames
you're going to want to actually collate together in the end.

If you do want to collate all, it checks to make sure that all of the
parameters that can be specified by a search are actually present in the
files it's adding to be collating. Additional code is added in
anticipation of having #5 done, which will hopefully ensure that any
duplicate batches are handled adequately.