Fix #6, Close #4, properly set "included_batch_files"
Forgot that even duplicate files should be in the included batch files,
to avoid accidentally including them later. So now the duplicate files
are also returned, so that they can be added to the
"included_batch_files" for the collated batch.

Also implemented code for #4 accidentally in the middle of this --
don't code in the middle of the night kids. I'm rusty with the coding +
closing issues thing lol. But in any case, it should collate all now
if you use that option.