Close #11, add last_updated attribute
Adds a last_updated attribute to the fandom_info.json and
fandom_names.json files that are saved by fandom_scrape.

I also had to comment out an import line in scrape.py because there was
some circular importing going on between fandom_scrape.py and scrape.py,
which was the reason for creating #14, which I had been meaning to
create anyway, but I have an even better reason now to.