Fix #32, Fix #25
Prevent it from showing all of the soup when there's no next URL if
there's a lot of soup. Only shows the first 50 characters and the last
50 characters.

The majority of #25 is also fixed here, so I'm going to mark it fixed.
It now successfully detects fics written by anonymous authors and
parses them successfully.