Close #26, find and rescrape anonymous works
Adds find_anon_entries.py, which looks for works that had anonymous
authors and weren't scraped properly, and then re-scrapes for them. The
newly scraped works are output in a json file beginning with
batch_anons_.

It does this by creating a search for the exact parameters of the file.
If the search brings up more than two entries, though, it labels it as
"ambiguous" and then outputs all the ambiguous works (and the urls that
were generated from them) into a json, whose name begins with
"stubborn_anons_".

find_anon_entries.py can also take two files containing ambiguous anons
and intersect them, finding all of the anons that were ambiguous both
times. (Unfortunately, I have yet to find a precise formula for really
finding all of the anons in one go, and with the same exact information
searched for, so it has to be played around with a little bit. I think
though that what I've got right now is pretty helpful, and gets at least
the majority of the anonymous works I had been looking at.)

I learned through this that you can't actually search for works that
have 0 words directly. Instead, you have to do a search for works with
0-1 words. Otherwise, you just get a search with any number of words.