Close #39, Fix #40 and fix #41
Add --remove-anons option for batch_utils, so that it removes anonymous
entries.

Fix the fact that it was not actually referencing files that were
implicitly included (by including other files, that were already batched
from those other other files).

Fix the bug where it was not properly removing duplicate works.

Discover #42, which is a Big Problem where it only got about 7/10 of all
the works it was supposed to scrape, which is very weird.