Fix #37, change default sort column
Change default sort column used in ao3_work_search_url to created_at
instead of _score, and changed the sort_direction to asc from desc. This
means that works will be shown from oldest created to newest created,
and creation date (as far as I know) is completely static. So now the
only problem will be if works are deleted. But hopefully this should be
less problematic for when it has to go back through after some of the
urls fail.

Plus, the ordering should at least be stable. Which is nice.