Close #43 and fix #44
Add code to find duplicates and only keep the one that was most recently
updated. (Not completely tested, to be honest)

I found out the hard way that it doesn't appropriately quit out of the
loop if you just want to not do anything when you find out that your
batch name is going to be a duplicate.