Close #48, handle KeyboardInterrupt
Gracefully quits out now, at least if you're in the try-catch block.
There needs to be more refactoring done there, but this should at least
make it so that I don't lose data that it's already collected.

It also provides helpful info, such as which page number was last
attempted (if applicable), and the number of works scarped. (Though that
might not be entirely necessary, since it prints that anyway.)