Further investigate #49
Not exactly sure what happened here. At all. I also had it go through to
see if there were any bad works in there that shouldn't have been in
there -- but there were only 2. I'm somewhat convinced that because
apparently the time in the linux machine was slightly ahead, it was able
to scrape more.

**By bad works, I mean works whose rating/category/etc. does not fit
what I actually wanted to search for.