Basically fix #57
Introduces proper checks to makes sure that urls that fail at the
beginning are handled. (This would all be handled anyway if I hadn't
accidentally introduced a base case that didn't need to be there in the
code, which is basically an exact copy of the loop before the loop even
begins.

This is further motivation to actually do something about #16.